refactor: split window_manager.rs into focused modules for better organization
Phase 3 refactoring complete - transformed 643-line monolithic window_manager.rs:

**New modular structure:**
- src/window_manager/mod.rs - Module interface and preserved tests (244 lines)
- src/window_manager/core.rs - WindowManager struct, initialization, main loop (91 lines)
- src/window_manager/events.rs - X11 event handling and dispatching (135 lines)
- src/window_manager/focus.rs - Focus management and visual borders (79 lines)
- src/window_manager/window_ops.rs - Layout integration and window operations (35 lines)

**Benefits achieved:**
- **Separation of concerns**: Event handling, focus, operations in focused modules
- **Maintainability**: 643 lines → 584 lines across focused files
- **Testability**: All 11 window manager tests preserved and passing
- **Readability**: Clear module boundaries and responsibilities
- **Extensibility**: Easy to add new event handlers or operations

**Code quality verified:**
- ✅ All 49 tests pass (no regressions)
- ✅ Cargo clippy clean
- ✅ Same public API maintained
- ✅ Added missing bsp_split_ratio() accessor to Config

Major refactoring milestone: 1039-line layout.rs + 643-line window_manager.rs
successfully modularized with comprehensive test coverage protection.
